users/serializers.py
from users.models import User
from rest_framework import serializers

class UpdateUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'phone_number', 'first_name', 'last_name', 'email')