requirements.txt
asgiref==3.5.2
certifi==2022.5.18
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.4
Django==3.2
django-cors-headers==3.12.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.5.0
djoser==2.0.3
filelock==3.7.1
fonttools==4.33.3
idna==3.3
importlib-metadata==4.11.3
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
mysql-connector-python==8.0.29
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.1
platformdirs==2.5.2
protobuf==4.21.1
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==2.0.0
social-auth-core==1.7.0
sqlparse==0.4.2
stripe==3.0.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
zipp==3.8.0
